refactor(py): genkit.core.reflection
CHANGELOG:
- [ ] Move reflection module into `genkit.core.reflection` as per
  request.
