+++ py/packages/genkit-ai/src/genkit/core/reflection.py
