+++ py/packages/genkit-ai/tests/genkit/core/endpoints/reflection_test.py
@@ -42,7 +42,7 @@ import pytest  # type: ignore
 import pytest_asyncio  # type: ignore
 from httpx import ASGITransport, AsyncClient  # type: ignore
 
-from genkit.core.endpoints.reflection import (
+from genkit.core.reflection import (
     create_reflection_asgi_app,  # type: ignore
 )
 from genkit.core.registry import Registry  # type: ignore
@@ -162,7 +162,7 @@ async def test_run_action_with_context(asgi_client, mock_registry):
 
 
 @pytest.mark.asyncio
-@patch('genkit.core.endpoints.reflection.is_streaming_requested')
+@patch('genkit.core.reflection.is_streaming_requested')
 async def test_run_action_streaming(
     mock_is_streaming, asgi_client, mock_registry
 ):
